move builder/testing to acctest inside sdk
