+++ builder/osc/bsusurrogate/builder_acc_test.go
@@ -3,7 +3,7 @@ package bsusurrogate
 import (
 	"testing"
 
-	builderT "github.com/hashicorp/packer/packer-plugin-sdk/builder/testing"
+	builderT "github.com/hashicorp/packer/packer-plugin-sdk/acctest"
 )
 
 func TestBuilderAcc_basic(t *testing.T) {
